first step to remove platforms
