+++ lib/page-object.rb
@@ -11,7 +11,6 @@ require 'page-object/section_collection'
 require 'page-object/widgets'
 
 require 'watir'
-require 'page-object/platforms/watir/element'
 require 'page-object/platforms/watir/page_object'
 
 require 'selenium/webdriver/common/error'
