+++ lib/page-object/elements/ordered_list.rb
@@ -4,11 +4,6 @@ module PageObject
     class OrderedList < Element
       include Enumerable
 
-      def initialize(element, platform)
-        @element = element
-        include_platform_for platform
-      end
-
       #
       # iterator that yields with a PageObject::Elements::ListItem
       #
@@ -20,7 +15,37 @@ module PageObject
         end
       end
 
+      #
+      # Return the PageObject::Elements::ListItem for the index provided.  Index
+      # is zero based.
+      #
+      # @return [PageObject::Elements::ListItem]
+      #
+      def [](idx)
+        Object::PageObject::Elements::ListItem.new(children[idx], :platform => :watir)
+      end
+
+      #
+      # Return the number of items contained in the ordered list
+      #
+      def items
+        children.size
+      end
+
+      #
+      # Return the ListItem objects that are children of the OrderedList
+      #
+      def list_items
+        children.collect do |obj|
+          Object::PageObject::Elements::ListItem.new(obj, :platform => :watir)
+        end
+      end
+
       protected
+
+      def children
+        element.ols(:xpath => child_xpath)
+      end
 
       def child_xpath
         "./child::li"
