+++ lib/page-object/elements/text_area.rb
@@ -3,17 +3,26 @@ module PageObject
   module Elements
     class TextArea < Element
 
-      def initialize(element, platform)
-        @element = element
-        include_platform_for platform
-      end
-
       def self.watir_finders
         super + [:label]
       end
 
       def self.selenium_finders
         super + [:label]
+      end
+
+      #
+      # Set the value of the TextArea
+      #
+      def value=(new_value)
+        element.set(new_value)
+      end
+
+      #
+      # Clear the TextArea
+      #
+      def clear
+        element.wd.clear
       end
 
     end
