--- lib/page-object/platforms/watir/file_field.rb
@@ -1,16 +1,0 @@
-module PageObject
-  module Platforms
-    module WatirWebDriver
-
-      module FileField
-        
-        #
-        # Set the value of the FileField
-        #
-        def value=(new_value)
-          element.set(new_value)
-        end
-      end
-    end
-  end
-end
