--- lib/page-object/platforms/watir/form.rb
@@ -1,16 +1,0 @@
-module PageObject
-  module Platforms
-    module WatirWebDriver
-      module Form
-
-        #
-        # Submit the form.
-        #
-        def submit
-          element.submit
-        end
-
-      end
-    end
-  end
-end
