--- lib/page-object/platforms/watir/link.rb
@@ -1,15 +1,0 @@
-module PageObject
-  module Platforms
-    module WatirWebDriver
-      module Link
-
-        #
-        # return the href for the link
-        #
-        def href
-          element.href
-        end
-      end
-    end
-  end
-end
