--- lib/page-object/platforms/watir/text_field.rb
@@ -1,16 +1,0 @@
-module PageObject
-  module Platforms
-    module WatirWebDriver
-
-      module TextField
-        
-        #
-        # Set the value of the TextField
-        #
-        def value=(new_value)
-          element.set(new_value)
-        end
-      end
-    end
-  end
-end
